refactor(cli): move ItemPicker and SearchOutputFormatter to util package
