+++ cli/src/main/kotlin/com/github/adriianh/cli/command/player/SearchCommand.kt
@@ -1,6 +1,8 @@
 package com.github.adriianh.cli.command.player
 
 import com.github.adriianh.cli.command.player.handler.SearchActionHandler
+import com.github.adriianh.cli.command.player.util.ItemPicker
+import com.github.adriianh.cli.command.player.util.SearchOutputFormatter
 import com.github.adriianh.cli.config.Messages
 import com.github.adriianh.cli.di.appModule
 import com.github.adriianh.core.domain.model.Track
